backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
pymongo==4.4.1
python-dotenv==1.0.0
Werkzeug==2.3.6
